# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Interview Position Tracker API",
    "description": "\n    A comprehensive REST API for tracking job positions and interview progress.\n    \n    ## Features\n    \n    * **User Authentication**: Secure JWT-based authentication system\n    * **Position Management**: Create, read, update, and delete job positions\n    * **Interview Tracking**: Track multiple interview stages for each position\n    * **Statistics & Analytics**: Get insights into your job search progress\n    * **Data Filtering**: Filter positions by status, company, date range, and more\n    * **Secure Access**: User-specific data isolation and authorization\n    \n    ## Authentication\n    \n    Most endpoints require authentication. To authenticate:\n    \n    1. Register a new account using `/api/v1/auth/register`\n    2. Login using `/api/v1/auth/login` to get an access token\n    3. Include the token in the Authorization header: `Bearer <your_token>`\n    \n    ## Workflow\n    \n    1. **Register/Login** to get authenticated\n    2. **Create positions** for jobs you're applying to\n    3. **Add interviews** as you progress through hiring processes\n    4. **Update statuses** as your applications move forward\n    5. **View statistics** to analyze your job search performance\n    ",
    "contact": {
      "name": "Interview Position Tracker API",
      "url": "https://github.com/your-repo/interview-tracker"
    },
    "license": {
      "name": "MIT",
...
  },
  "paths": {
...
    "/api/v1/statistics/overview": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Get Overview Statistics",
        "description": "Get overview statistics for the current user's job applications.\n\nThis endpoint provides comprehensive metrics including:\n- Total applications, companies, and interviews\n- Response rate, interview rate, and offer rate\n- Breakdowns by status, interview type, and interview outcome\n\nOptional filters can be applied to focus on specific time periods,\ncompanies, or application statuses.",
        "operationId": "get_overview_statistics_api_v1_statistics_overview_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter start date",
              "title": "Start Date"
            },
            "description": "Filter start date"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter end date",
              "title": "End Date"
            },
            "description": "Filter end date"
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by company name",
              "title": "Company"
            },
            "description": "Filter by company name"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PositionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by position status",
              "title": "Status"
            },
            "description": "Filter by position status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsOverview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/statistics/timeline": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Get Timeline Statistics",
        "description": "Get timeline-based statistics for the current user's job applications.\n\nThis endpoint provides time-based metrics including:\n- Applications and interviews per month\n- Average response time from application to first interview\n- Average time from interview to final decision\n\nIf no date range is specified, the endpoint will use the full range\nof the user's application dates.",
        "operationId": "get_timeline_statistics_api_v1_statistics_timeline_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter start date",
              "title": "Start Date"
            },
            "description": "Filter start date"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter end date",
              "title": "End Date"
            },
            "description": "Filter end date"
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by company name",
              "title": "Company"
            },
            "description": "Filter by company name"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PositionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by position status",
              "title": "Status"
            },
            "description": "Filter by position status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/statistics/companies": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Get Company Statistics",
        "description": "Get company-based statistics for the current user's job applications.\n\nThis endpoint provides per-company metrics including:\n- Total applications and interviews per company\n- Latest application date for each company\n- Status breakdown for each company\n- Success rate (offers/applications) for each company\n\nCompanies are sorted by total applications in descending order.",
        "operationId": "get_company_statistics_api_v1_statistics_companies_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter start date",
              "title": "Start Date"
            },
            "description": "Filter start date"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter end date",
              "title": "End Date"
            },
            "description": "Filter end date"
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by company name",
              "title": "Company"
            },
            "description": "Filter by company name"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PositionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by position status",
              "title": "Status"
            },
            "description": "Filter by position status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyStatisticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompanyStatistics": {
        "properties": {
          "company_name": {
            "type": "string",
            "title": "Company Name",
            "description": "Company name"
          },
          "total_applications": {
            "type": "integer",
            "title": "Total Applications",
            "description": "Total applications to this company"
          },
          "total_interviews": {
            "type": "integer",
            "title": "Total Interviews",
            "description": "Total interviews with this company"
          },
          "latest_application_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Application Date",
            "description": "Most recent application date"
          },
          "status_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/PositionStatus"
            },
            "type": "object",
            "title": "Status Breakdown",
            "description": "Status breakdown for this company"
          },
          "success_rate": {
            "type": "number",
            "title": "Success Rate",
            "description": "Success rate (offers/applications)"
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "total_applications",
          "total_interviews",
          "status_breakdown",
          "success_rate"
        ],
        "title": "CompanyStatistics",
        "description": "Schema for company-based statistics."
      },
      "CompanyStatisticsResponse": {
        "properties": {
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyStatistics"
            },
            "type": "array",
            "title": "Companies",
            "description": "Statistics per company"
          },
          "total_companies": {
            "type": "integer",
            "title": "Total Companies",
            "description": "Total number of companies"
          }
        },
        "type": "object",
        "required": [
          "companies",
          "total_companies"
        ],
        "title": "CompanyStatisticsResponse",
        "description": "Schema for company statistics response."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterviewOutcome": {
        "type": "string",
        "enum": [
          "pending",
          "passed",
          "failed",
          "cancelled"
        ],
        "title": "InterviewOutcome",
        "description": "Outcome of an interview."
      },
...
      "InterviewType": {
        "type": "string",
        "enum": [
          "technical",
          "behavioral",
          "hr",
          "final"
        ],
        "title": "InterviewType",
        "description": "Type of interview based on its purpose."
      },
...
      "PositionStatus": {
        "type": "string",
        "enum": [
          "applied",
          "screening",
          "interviewing",
          "offer",
          "rejected",
          "withdrawn"
        ],
        "title": "PositionStatus",
        "description": "Status of a job position application."
      },
...
      "StatisticsOverview": {
        "properties": {
          "total_applications": {
            "type": "integer",
            "title": "Total Applications",
            "description": "Total number of applications"
          },
          "total_companies": {
            "type": "integer",
            "title": "Total Companies",
            "description": "Total number of unique companies"
          },
          "total_interviews": {
            "type": "integer",
            "title": "Total Interviews",
            "description": "Total number of interviews"
          },
          "response_rate": {
            "type": "number",
            "title": "Response Rate",
            "description": "Percentage of applications that got responses"
          },
          "interview_rate": {
            "type": "number",
            "title": "Interview Rate",
            "description": "Percentage of applications that led to interviews"
          },
          "offer_rate": {
            "type": "number",
            "title": "Offer Rate",
            "description": "Percentage of applications that led to offers"
          },
          "status_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/PositionStatus"
            },
            "type": "object",
            "title": "Status Breakdown",
            "description": "Count by position status"
          },
          "interview_type_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/InterviewType"
            },
            "type": "object",
            "title": "Interview Type Breakdown",
            "description": "Count by interview type"
          },
          "interview_outcome_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/InterviewOutcome"
            },
            "type": "object",
            "title": "Interview Outcome Breakdown",
            "description": "Count by interview outcome"
          }
        },
        "type": "object",
        "required": [
          "total_applications",
          "total_companies",
          "total_interviews",
          "response_rate",
          "interview_rate",
          "offer_rate",
          "status_breakdown",
          "interview_type_breakdown",
          "interview_outcome_breakdown"
        ],
        "title": "StatisticsOverview",
        "description": "Schema for general statistics overview."
      },
      "TimelineStatistics": {
        "properties": {
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start",
            "description": "Start date of the period"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End",
            "description": "End date of the period"
          },
          "applications_per_month": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "object"
            },
            "type": "array",
            "title": "Applications Per Month",
            "description": "Applications count per month"
          },
          "interviews_per_month": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "object"
            },
            "type": "array",
            "title": "Interviews Per Month",
            "description": "Interviews count per month"
          },
          "average_response_time_days": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Response Time Days",
            "description": "Average days to get response"
          },
          "average_interview_to_decision_days": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Interview To Decision Days",
            "description": "Average days from interview to decision"
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "period_end",
          "applications_per_month",
          "interviews_per_month"
        ],
        "title": "TimelineStatistics",
        "description": "Schema for time-based statistics."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}